plantAPI/requirements.txt
Pillow==9.0.1
torch==1.10.1+cu111
torchvision==0.11.2+cu111